Move test that check questions' length from ExamController to Exam
