+++ apps/hardhat/contracts/Exam.sol
@@ -25,9 +25,24 @@ contract Exam {
 
     function addQuestion(Question calldata _question) public {
         require(
-            _question.answers.length < 7,
+            _question.answers.length <= 6,
             "The maximum number of answers you can add to a single question is 6."
         );
+
+        bool isTheSameId = false;
+
+        for (uint256 i = 0; i < questions.length; i++) {
+            if (questions[i].id == _question.id) {
+                isTheSameId = true;
+                break;
+            }
+        }
+
+        require(
+            !isTheSameId,
+            "Some of the questions you have provided have the same id."
+        );
+
         questions.push(_question);
     }
