+++ apps/hardhat/test/ExamController.test.ts
@@ -90,35 +90,4 @@ describe('ExamController tests', () => {
       examController.addExam(EXAM_NAME, EXAM_DESCRIPTION, tooManyQuestions)
     ).to.be.revertedWith('The maximum number of questions you can add is 30.');
   });
-
-  it("Shouldn't add Exam with more than 6 answers to a single question", async () => {
-    const { examController } = await loadFixture(deployExamControllerFixture);
-
-    const basicAnswer = MOCK_ANSWERS.at(0)!;
-
-    const answersArray = new Array(7).fill(0).map((e, idx) => ({
-      ...basicAnswer,
-      id: idx,
-    }));
-
-    const questionsWithTooManyAnswers: Question[] = [
-      ...MOCK_QUESTIONS,
-      {
-        id: 2,
-        description: EXAM_DESCRIPTION,
-        header: EXAM_NAME,
-        answers: answersArray,
-      },
-    ];
-
-    await expect(
-      examController.addExam(
-        EXAM_NAME,
-        EXAM_DESCRIPTION,
-        questionsWithTooManyAnswers
-      )
-    ).to.be.revertedWith(
-      'The maximum number of answers you can add to a single question is 6.'
-    );
-  });
 });
